34 | add like button with fontawesome

+++ articles/views.py
@@ -8,7 +8,7 @@ from .models import Article, Comment
 # from accounts.models import User
 from django.contrib.auth import get_user_model
 from django.conf import settings
-from django.http import HttpResponseForbidden
+from django.http import HttpResponseForbidden, HttpResponse
 
 # Create your views here.
 @require_GET
@@ -76,16 +76,18 @@ def detail(request, article_pk):
     return render(request, 'articles/detail.html', context)
     
 @require_POST
-@login_required
 def delete(request, article_pk):
-    # article = Article.objects.get(pk=article_pk)
-    article = get_object_or_404(Article, pk=article_pk)
-    if article.user == request.user:
-    # if request.method == "POST":
-        article.delete()
-        return redirect('articles:index')
+    if request.user.is_authenticated:
+        # article = Article.objects.get(pk=article_pk)
+        article = get_object_or_404(Article, pk=article_pk)
+        if article.user == request.user:
+        # if request.method == "POST":
+            article.delete()
+            return redirect('articles:index')
+        else:
+            return HttpResponseForbidden()
     else:
-        return HttpResponseForbidden()
+        return HttpResponse('Unauthorized', status=401)
 
 # def edit(request, article_pk):
 #     if request.method == 'GET':
@@ -124,24 +126,26 @@ def update(request, article_pk):
         return HttpResponseForbidden()
 
 @require_POST
-@login_required
 def comment_create(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    # 1. modelform에 사용자 입력값 넣고
-    comment_form = CommentForm(request.POST)
-    # 2. 검증하고,
-    if comment_form.is_valid():
-    # 3. 맞으면 저장!
-    # 3-1. 사용자 입력값으로 comment instance 생성 (저장은 X)
-        comment = comment_form.save(commit=False)
-        # 3-2. FK 넣고 저장
-        comment.article = article
-        comment.user = request.user
-        comment.save()
-    # 4. return redirect
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        # 1. modelform에 사용자 입력값 넣고
+        comment_form = CommentForm(request.POST)
+        # 2. 검증하고,
+        if comment_form.is_valid():
+        # 3. 맞으면 저장!
+        # 3-1. 사용자 입력값으로 comment instance 생성 (저장은 X)
+            comment = comment_form.save(commit=False)
+            # 3-2. FK 넣고 저장
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+        # 4. return redirect
+        else:
+            messages.info(request, '댓글 형식이 맞지 않습니다.')
+        return redirect('articles:detail', article_pk)
     else:
-        messages.info(request, '댓글 형식이 맞지 않습니다.')
-    return redirect('articles:detail', article_pk)
+        return HttpResponse('Unauthorized', status=401)
 
     # comment = Comment()
     # comment.content = request.POST.get('content')
